app.py
from flask import Flask, render_template, request, redirect
import mysql.connector
import bcrypt

app = Flask(__name__)

def connect_db():
    return mysql.connector.connect(
        host="localhost",
        user="root",
        password="123456",
        database="file_sharing"
    )

@app.route('/')
def home():
    return "Hello, this is my first file sharing app!"

@app.route('/register', methods=['GET', 'POST'])
def register():
    if request.method == 'POST':
        username = request.form['username']
        password = request.form['password'].encode('utf-8')
        password_hash = bcrypt.hashpw(password, bcrypt.gensalt())
        try:
            db = connect_db()
            cursor = db.cursor()
            cursor.execute(
                "INSERT INTO users (username, password_hash) VALUES (%s, %s)",
                (username, password_hash)
            )
            db.commit()
            cursor.close()
            db.close()
            return redirect('/')
        except mysql.connector.Error as e:
            return f"Error: {e}"
    return render_template('register.html')

if __name__ == '__main__':
    app.run(host='0.0.0.0', port=5001, debug=True)